friend 로직 user에서 분리 및 aop common으로 이동 그 외 리뷰사항 반영
