+++ src/main/java/com/dlrtn/websocket/chat/business/friend/api/FriendApiController.java
@@ -1,12 +1,12 @@
-package com.dlrtn.websocket.chat.business.user.api;
+package com.dlrtn.websocket.chat.business.friend.api;
 
-import com.dlrtn.websocket.chat.business.user.aop.SessionUser;
-import com.dlrtn.websocket.chat.business.user.application.FriendService;
+import com.dlrtn.websocket.chat.business.friend.application.FriendService;
+import com.dlrtn.websocket.chat.business.friend.model.payload.AddFriendResponse;
+import com.dlrtn.websocket.chat.business.friend.model.payload.ChangeFriendStateRequest;
+import com.dlrtn.websocket.chat.business.friend.model.payload.ChangeFriendStateResponse;
+import com.dlrtn.websocket.chat.business.friend.model.payload.DeleteFriendResponse;
 import com.dlrtn.websocket.chat.business.user.model.domain.User;
-import com.dlrtn.websocket.chat.business.user.model.payload.AddFriendResponse;
-import com.dlrtn.websocket.chat.business.user.model.payload.ChangeFriendStateRequest;
-import com.dlrtn.websocket.chat.business.user.model.payload.ChangeFriendStateResponse;
-import com.dlrtn.websocket.chat.business.user.model.payload.DeleteFriendResponse;
+import com.dlrtn.websocket.chat.common.aop.SessionUser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -18,7 +18,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
-public class UserFriendController {
+public class FriendApiController {
 
     private final FriendService friendService;
